final-pjt-back/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.3
django-bootstrap-v5==1.0.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
importlib-metadata==3.10.1
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
zipp==3.4.1
